context: Print error and gracefully handle NoDataAbortContent Exception
We're seeing on some ARM platforms that our attempts to crash a
platform, thereby triggering a Data Abort handler, does not work.
Instead, it seems that the target just returns a rotated word or
garbage. (See Tetrel-Depthcharge #5)

The best we can do for now is to just show the error and move on.